Code Changes and added Authentication Page
