+++ src/App.js
@@ -1,12 +1,10 @@
 import {
-  BrowserRouter as Router,
-  Switch,
   Routes,
   Route
 } from "react-router-dom";
 
 import LoginForm from "./Components/LoginForm/LoginForm.js";
-import HomePage from "./Components/HomePage/HomePage.js";
+import Authentication from "./Components/Authentication.js";
 
 function App() {
 
@@ -14,8 +12,9 @@ function App() {
     <>
     <Routes>
        <Route exact path='/' element={<LoginForm/>}/>
-       <Route exact path='/Home' element={<HomePage pageName='Create'/>}/>
-       <Route exact path='/Search' element={<HomePage pageName='Search'/>}/>
+       <Route exact path='/Home' element={<Authentication pageName='HomePage'/>}/>
+       <Route exact path='/Search' element={<Authentication pageName='Search'/>}/>
+       <Route path="*" element={<Authentication/>} />
        </Routes>
     </>
   );
